- Перераспределил файлы; - Переписал верстку кодом.
