+++ Tracker/Trackers/TrackersViewController.swift
@@ -1,0 +1,125 @@
+//
+//  TrackersViewController.swift
+//  Tracker
+//
+//  Created by Вадим on 29.07.2024.
+//
+
+import UIKit
+
+final class TrackersViewController: UIViewController {
+    
+    // MARK: - Private Properties
+    
+    private lazy var addButton: UIBarButtonItem = {
+        let button = UIBarButtonItem(
+            image: UIImage(named: "Plus"),
+            style: .plain,
+            target: self,
+            action: #selector(addTracker))
+        button.tintColor = .black
+        return button
+    }()
+    
+    private lazy var datePicker: UIBarButtonItem = {
+        let datePicker = UIDatePicker()
+        datePicker.datePickerMode = .date
+        datePicker.addTarget(
+            self,
+            action: #selector(datePickerValueChanged(_:)),
+            for: .valueChanged)
+        return UIBarButtonItem(customView: datePicker)
+    }()
+    
+    private let titleLabel: UILabel = {
+        let label = UILabel()
+        label.text = "Трекеры"
+        label.font = UIFont.systemFont(ofSize: 34, weight: .bold)
+        return label
+    }()
+    
+    private let searchBar: UISearchBar = {
+        let searchBar = UISearchBar()
+        searchBar.placeholder = "Поиск"
+        return searchBar
+    }()
+    
+    private let errorImageView: UIImageView = {
+        let imageView = UIImageView(image: UIImage(named: "Error"))
+        return imageView
+    }()
+    
+    private let trackingLabel: UILabel = {
+        let label = UILabel()
+        label.text = "Что будем отслеживать?"
+        label.font = UIFont.systemFont(ofSize: 12, weight: .medium)
+        label.textAlignment = .center
+        return label
+    }()
+    
+    // MARK: - Lifecycle
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .white
+        setupNavigationBar()
+        setupUI()
+    }
+    
+    override var preferredStatusBarStyle: UIStatusBarStyle {
+        return .darkContent
+    }
+    
+    // MARK: - Private Methods
+    
+    private func setupNavigationBar() {
+        navigationItem.leftBarButtonItems = [addButton]
+        navigationItem.rightBarButtonItems = [datePicker]
+    }
+    
+    private func setupUI() {
+        [titleLabel,
+         searchBar,
+         errorImageView,
+         trackingLabel].forEach { [weak self] view in
+            guard let self else { return }
+            self.view.addSubview(view)
+            view.translatesAutoresizingMaskIntoConstraints = false
+        }
+        
+        NSLayoutConstraint.activate([
+            titleLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 16),
+            titleLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
+            titleLabel.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
+            
+            searchBar.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 7),
+            searchBar.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
+            searchBar.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
+            searchBar.heightAnchor.constraint(equalToConstant: 36),
+            
+            errorImageView.widthAnchor.constraint(equalToConstant: 80),
+            errorImageView.heightAnchor.constraint(equalToConstant: 80),
+            errorImageView.topAnchor.constraint(equalTo: view.topAnchor, constant: 402),
+            errorImageView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            
+            trackingLabel.topAnchor.constraint(equalTo: errorImageView.bottomAnchor, constant: 8),
+            trackingLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            trackingLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
+            trackingLabel.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
+            trackingLabel.heightAnchor.constraint(equalToConstant: 18)
+        ])
+    }
+    
+    // MARK: - Actions
+    
+    @objc private func addTracker() {
+        print("Добавить трекер")
+    }
+    
+    @objc private func datePickerValueChanged(_ sender: UIDatePicker) {
+        let datePicker = DateFormatter()
+        datePicker.dateFormat = "dd.MM.yy"
+//        let dateString = datePicker.string(from: sender.date)
+//        print("Дата изменена на \(dateString)")
+    }
+}
